+++ packages/http/src/payload.ts
@@ -361,7 +361,7 @@ function resolveMultiPartBodyFromModel(
   const diagnostics = createDiagnosticCollector();
   const parts: HttpOperationPart[] = [];
   for (const item of type.properties.values()) {
-    const part = diagnostics.pipe(resolvePartOrParts(program, item.type, visibility));
+    const part = diagnostics.pipe(resolvePartOrParts(program, item.type, visibility, item));
     if (part) {
       parts.push({ ...part, name: part.name ?? item.name, optional: item.optional });
     }
@@ -441,15 +441,16 @@ function resolvePartOrParts(
   program: Program,
   type: Type,
   visibility: Visibility,
+  property?: ModelProperty,
 ): DiagnosticResult<HttpOperationPart | undefined> {
   if (type.kind === "Model" && isArrayModelType(program, type)) {
-    const [part, diagnostics] = resolvePart(program, type.indexer.value, visibility);
+    const [part, diagnostics] = resolvePart(program, type.indexer.value, visibility, property);
     if (part) {
       return [{ ...part, multi: true }, diagnostics];
     }
     return [part, diagnostics];
   } else {
-    return resolvePart(program, type, visibility);
+    return resolvePart(program, type, visibility, property);
   }
 }
 
@@ -457,6 +458,7 @@ function resolvePart(
   program: Program,
   type: Type,
   visibility: Visibility,
+  property?: ModelProperty,
 ): DiagnosticResult<HttpOperationPart | undefined> {
   const diagnostics = createDiagnosticCollector();
   const part = getHttpPart(program, type);
@@ -489,6 +491,7 @@ function resolvePart(
 
     return diagnostics.wrap({
       multi: false,
+      property,
       name: part.options.name,
       body,
       optional: false,