+++ packages/spector/src/app/app.ts
@@ -1,4 +1,11 @@
-import { MockApiDefinition, MockRequest, RequestExt, ScenarioMockApi } from "@typespec/spec-api";
+import {
+  MockApiDefinition,
+  MockBody,
+  MockMultipartBody,
+  MockRequest,
+  RequestExt,
+  ScenarioMockApi,
+} from "@typespec/spec-api";
 import { ScenariosMetadata } from "@typespec/spec-coverage-sdk";
 import { Response, Router } from "express";
 import { getScenarioMetadata } from "../coverage/common.js";
@@ -59,73 +66,54 @@ export class MockApiApp {
 
   private registerScenario(name: string, scenario: ScenarioMockApi) {
     for (const endpoint of scenario.apis) {
-      if (endpoint.kind !== "MockApiDefinition") {
-        this.router.route(endpoint.uri)[endpoint.method]((req: RequestExt, res: Response) => {
-          processRequest(
-            this.coverageTracker,
-            name,
-            endpoint.uri,
-            req,
-            res,
-            endpoint.handler,
-          ).catch((e) => {
+      if (!endpoint.handler) {
+        endpoint.handler = createHandler(endpoint);
+      }
+      this.router.route(endpoint.uri)[endpoint.method]((req: RequestExt, res: Response) => {
+        processRequest(this.coverageTracker, name, endpoint.uri, req, res, endpoint.handler!).catch(
+          (e) => {
             logger.error("Unexpected request error", e);
             res.status(500).end();
-          });
-        });
-      } else {
-        if (!endpoint.handler) {
-          endpoint.handler = createHandler(endpoint);
-        }
-        this.router.route(endpoint.uri)[endpoint.method]((req: RequestExt, res: Response) => {
-          processRequest(
-            this.coverageTracker,
-            name,
-            endpoint.uri,
-            req,
-            res,
-            endpoint.handler!,
-          ).catch((e) => {
-            logger.error("Unexpected request error", e);
-            res.status(500).end();
-          });
-        });
+          },
+        );
+      });
+    }
+  }
+}
+
+function validateBody(req: MockRequest, body: MockBody | MockMultipartBody) {
+  if ("kind" in body) {
+    // custom handler for now.
+  } else {
+    if (Buffer.isBuffer(body.rawContent)) {
+      req.expect.rawBodyEquals(body.rawContent);
+    } else {
+      switch (body.contentType) {
+        case "application/json":
+          req.expect.coercedBodyEquals(JSON.parse(body.rawContent as any));
+          break;
+        case "application/xml":
+          req.expect.xmlBodyEquals(
+            (body.rawContent as any).replace(`<?xml version='1.0' encoding='UTF-8'?>`, ""),
+          );
+          break;
+        default:
+          req.expect.rawBodyEquals(body.rawContent);
       }
     }
   }
 }
 
-function isObject(value: any): boolean {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
 function createHandler(apiDefinition: MockApiDefinition) {
   return (req: MockRequest) => {
+    const body = apiDefinition.request?.body;
     // Validate body if present in the request
-    if (apiDefinition.request.body) {
-      if (
-        apiDefinition.request.headers &&
-        apiDefinition.request.headers["Content-Type"] === "application/xml"
-      ) {
-        req.expect.xmlBodyEquals(
-          apiDefinition.request.body.rawContent.replace(
-            `<?xml version='1.0' encoding='UTF-8'?>`,
-            "",
-          ),
-        );
-      } else {
-        if (isObject(apiDefinition.request.body)) {
-          Object.entries(apiDefinition.request.body).forEach(([key, value]) => {
-            req.expect.deepEqual(req.body[key], value);
-          });
-        } else {
-          req.expect.coercedBodyEquals(apiDefinition.request.body);
-        }
-      }
+    if (body) {
+      validateBody(req, body);
     }
 
     // Validate headers if present in the request
-    if (apiDefinition.request.headers) {
+    if (apiDefinition.request?.headers) {
       Object.entries(apiDefinition.request.headers).forEach(([key, value]) => {
         if (key.toLowerCase() !== "content-type") {
           if (Array.isArray(value)) {
@@ -137,21 +125,12 @@ function createHandler(apiDefinition: MockApiDefinition) {
       });
     }
 
-    // Validate query params if present in the request
-    if (apiDefinition.request.params) {
-      Object.entries(apiDefinition.request.params).forEach(([key, value]) => {
-        if (!req.query[key]) {
-          if (Array.isArray(value)) {
-            req.expect.deepEqual(req.params[key], value);
-          } else {
-            req.expect.deepEqual(req.params[key], String(value));
-          }
+    if (apiDefinition.request?.query) {
+      Object.entries(apiDefinition.request.query).forEach(([key, value]) => {
+        if (Array.isArray(value)) {
+          req.expect.deepEqual(req.query[key], value);
         } else {
-          if (Array.isArray(value)) {
-            req.expect.deepEqual(req.query[key], value);
-          } else {
-            req.expect.containsQueryParam(key, String(value));
-          }
+          req.expect.containsQueryParam(key, String(value));
         }
       });
     }